get_acceleration {aws.s3}	R Documentation

Bucket Acceleration

Description

Get/Put acceleration settings or retrieve acceleration status of a bucket.

Usage

get_acceleration(bucket, ...)

put_acceleration(bucket, status = c("Enabled", "Suspended"), ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

status
Character string specifying whether acceleration should be “Enabled” or “Suspended”.

Details

Transfer acceleration is a AWS feature that enables potentially faster file transfers to and from S3, particularly when making cross-border transfers (such as from a European client location to the us-east-1 S3 region). Acceleration must be enabled before it can be used. Once enabled, accelerate = TRUE can be passed to any aws.s3 function via s3HTTP. get_acceleration returns the acceleration status of a bucket; put_acceleration enables or suspends acceleration.

Value

For get_acceleration: If acceleration has never been enabled or suspend, the value is NULL. Otherwise, the status is returned (either “Enabled” or “Suspended”). For put_acceleration: If acceleration has never been enabled or suspend, the value is NULL.

References

API Documentation: PUT Bucket accelerate API Documentation: GET Bucket accelerate

Examples

## Not run: 
b <- bucketlist()
get_acceleration(b[[1]])
put_acceleration(b[[1]], "Enabled")
get_acceleration(b[[1]])
put_acceleration(b[[1]], "Suspended")

## End(Not run)

[Package aws.s3 version 0.3.21 Index]