get_acl {aws.s3}	R Documentation

Get or put bucket/object ACLs

Description

Access Control Lists (ACLs) control access to buckets and objects. These functions retrieve and modify ACLs for either objects or buckets.

Usage

get_acl(object, bucket, ...)

put_acl(object, bucket, acl = NULL, headers = list(), body = NULL, ...)

Arguments

object
Character string with the object key, or an object of class “s3_object”. In most cases, if object is specified as the latter, bucket can be omitted because the bucket name will be extracted from “Bucket” slot in object.

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

acl
A character string indicating a “canned” access control list. By default all bucket contents and objects therein are given the ACL “private”. This can later be viewed using get_acl and modified using put_acl.

headers
List of request headers for the REST call

body
A character string containing an XML-formatted ACL.

Details

get_acl retrieves an XML-formatted ACL for either an object (if specified) or a bucket (if specified).

Value

For get_acl a character string containing an XML-formatted ACL. For put_acl: if successful, TRUE.

References

API Reference: GET Object ACL API Reference: PUT Object ACL

[Package aws.s3 version 0.3.21 Index]