bucketlist {aws.s3}	R Documentation

List Buckets

Description

List buckets as a data frame

Usage

bucketlist(add_region = FALSE, ...)

bucket_list_df(add_region = FALSE, ...)

Arguments

add_region
A logical (by default FALSE) indicating whether to add “Region” column to the output data frame. This simply induces a loop over get_location for each bucket.

...
Additional arguments passed to s3HTTP.

Details

bucketlist performs a GET operation on the base s3 endpoint and returns a list of all buckets owned by the authenticated sender of the request. If authentication is successful, this function provides a list of buckets available to the authenticated user. In this way, it can serve as a “hello world!” function, to confirm that one's authentication credentials are working correctly.

bucket_list_df and bucketlist are identical.

Value

A data frame of buckets. Can be empty (0 rows, 0 columns) if there are no buckets, otherwise contains typically at least columns Bucket and CreationDate.

References

API Documentation

See Also

get_bucket, get_object

[Package aws.s3 version 0.3.21 Index]