copy_object {aws.s3}	R Documentation

Copy Objects

Description

Copy objects between S3 buckets

Usage

copy_object(
  from_object,
  to_object = from_object,
  from_bucket,
  to_bucket,
  headers = list(),
  ...
)

copy_bucket(from_bucket, to_bucket, ...)

Arguments

from_object
A character string containing the name the object you want to copy.

to_object
A character string containing the name the object should have in the new bucket.

from_bucket
A character string containing the name of the bucket you want to copy from.

to_bucket
A character string containing the name of the bucket you want to copy into.

headers
List of request headers for the REST call.

...
Additional arguments passed to s3HTTP.

Details

copy_object copies an object from one bucket to another without bringing it into local memory. For copy_bucket, all objects from one bucket are copied to another (limit 1000 objects). The same keys are used in the old bucket as in the new bucket.

Value

Something...

References

API Documentation

[Package aws.s3 version 0.3.21 Index]