get_cors {aws.s3}	R Documentation

CORS

Description

Get/Put/Delete the cross origin resource sharing configuration information for a bucket.

Usage

get_cors(bucket, ...)

put_cors(bucket, ...)

delete_cors(bucket, ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

Value

For get_cors: A list with cors configuration and rules. For delete_cors: TRUE if successful, FALSE otherwise.

References

API Documentation: PUT cors API Documentation: GET cords API Documentation: DELETE cors

[Package aws.s3 version 0.3.21 Index]