get_location {aws.s3}	R Documentation

Bucket location

Description

Get the AWS region location of bucket.

Usage

get_location(bucket, ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

Value

A character string containing the region, if one has been set.

References

API Documentation

[Package aws.s3 version 0.3.21 Index]