get_lifecycle {aws.s3}	R Documentation

Lifecycle

Description

Get/Put/Delete the lifecycle configuration information for a bucket.

Usage

get_lifecycle(bucket, ...)

put_lifecycle(bucket, request_body, ...)

delete_lifecycle(bucket, ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

request_body
A character string containing an XML request body, as defined in the specification in the API Documentation.

Value

For get_lifecycle: a list with lifecycle configuration, if it has been configured. For delete_lifecycle: TRUE if successful, FALSE otherwise.

References

API Documentation: PUT lifecycle API Documentation: GET lifecycle API Documentation: DELETE lifecycle

[Package aws.s3 version 0.3.21 Index]