get_notification {aws.s3}	R Documentation

Notifications

Description

Get/put the notification configuration for a bucket.

Usage

get_notification(bucket, ...)

put_notification(bucket, request_body, ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

request_body
A character string containing an XML request body, as defined in the specification in the API Documentation.

Value

A list containing the notification configuration, if one has been set.

References

API Documentation: GET API Documentation: PUT

[Package aws.s3 version 0.3.21 Index]