s3saveRDS {aws.s3}	R Documentation

saveRDS/readRDS

Description

Serialization interface to read/write R objects to S3

Usage

s3saveRDS(
  x,
  object = paste0(as.character(substitute(x)), ".rds"),
  bucket,
  compress = TRUE,
  ...
)

s3readRDS(object, bucket, ...)

Arguments

x
For s3saveRDS, a single R object to be saved via saveRDS and uploaded to S3. x is analogous to the object argument in saveRDS.

object
Character string with the object key, or an object of class “s3_object”. In most cases, if object is specified as the latter, bucket can be omitted because the bucket name will be extracted from “Bucket” slot in object.

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

compress
A logical. See saveRDS.

...
Additional arguments passed to s3HTTP.

Details

Note that early versions of s3saveRDS from aws.s3 <= 0.2.4 unintentionally serialized objects to big endian format (due to defaults in serialize. This can create problems when attempting to read these files using readRDS. The function attempts to catch the issue and read accordingly, but may fail. The solution used internally is unserialize(memDecompress(get_object(), "gzip"))

Value

For s3saveRDS, a logical. For s3readRDS, an R object.

Author(s)

Steven Akins <skawesome@gmail.com>

See Also

s3save,s3load

Examples

## Not run: 
# create bucket
b <- put_bucket("myexamplebucket")

# save a single object to s3
s3saveRDS(x = mtcars, bucket = "myexamplebucket", object = "mtcars.rds")

# restore it under a different name
mtcars2 <- s3readRDS(object = "mtcars.rds", bucket = "myexamplebucket")
identical(mtcars, mtcars2)

# cleanup
delete_object(object = "mtcars.rds", bucket = "myexamplebucket")
delete_bucket("myexamplebucket")

## End(Not run)

[Package aws.s3 version 0.3.21 Index]