get_tagging {aws.s3}	R Documentation

Bucket tagging

Description

Get/delete the tag set for a bucket.

Usage

get_tagging(bucket, ...)

put_tagging(bucket, tags = list(), ...)

delete_tagging(bucket, ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

tags
A list containing key-value pairs of tag names and values.

Value

A list containing the tag set, if one has been set. For delete_tagging: TRUE if successful, FALSE otherwise.

References

API Documentation: PUT tagging API Documentation: GET tagging API Documentation: DELETE tagging

Examples

## Not run: 
 put_tagging("mybucket", tags = list(foo = "1", bar = "2"))
 get_tagging("mybucket")
 delete_tagging("mybucket")

## End(Not run)

[Package aws.s3 version 0.3.21 Index]