get_versions {aws.s3}	R Documentation

Bucket versions

Description

Get/Put versioning settings or retrieve versions of bucket objects.

Usage

get_versions(bucket, ...)

get_versioning(bucket, ...)

put_versioning(bucket, status = c("Enabled", "Suspended"), ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

status
Character string specifying whether versioning should be “Enabled” or “Suspended”.

Details

get_versioning returns the versioning status of a bucket; put_versioning sets the versioning status. get_versions returns information about bucket versions.

Value

For get_versioning: If versioning has never been enabled or suspend, the value is NULL. Otherwise, the status is returned (either “Enabled” or “Suspended”). For put_versioning: If versioning has never been enabled or suspend, the value is NULL. Otherwise, the status is returned (either “Enabled” or “Suspended”). For get_versions: A list.

References

API Documentation API Documentation API Documentation

Examples

## Not run: 
 put_versioning("mybucket")
 get_versioning("mybucket")
 get_versions("mybucket")

## End(Not run)

[Package aws.s3 version 0.3.21 Index]