delete_website {aws.s3}	R Documentation

Bucket Website configuration

Description

Get/Put/Delete the website configuration for a bucket.

Usage

delete_website(bucket, ...)

put_website(bucket, request_body, ...)

get_website(bucket, ...)

Arguments

bucket
Character string with the name of the bucket, or an object of class “s3_bucket”.

...
Additional arguments passed to s3HTTP.

request_body
A character string containing an XML request body, as defined in the specification in the API Documentation.

Value

For put_website and get_website, a list containing the website configuration, if one has been set. For delete_website: TRUE if successful, FALSE otherwise. An aws_error object may be returned if the request failed.

References

API Documentation: PUT website API Documentation: GET website API Documentation: DELETE website

[Package aws.s3 version 0.3.21 Index]